Обновил контроллеры и коллекции, переместил функционал из админок на обычный контроллер
