--- collections/login-reg.postman_collection.json
@@ -1,120 +1,0 @@
-{
-	"info": {
-		"_postman_id": "600aa90d-59b2-43ef-96e6-8e0794b70a59",
-		"name": "login/reg",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-		"_exporter_id": "40388756",
-		"_collection_link": "https://test99-0330.postman.co/workspace/test~7488d7e6-fe20-496e-ad1d-0a4b9129da6b/collection/40388756-600aa90d-59b2-43ef-96e6-8e0794b70a59?action=share&source=collection_link&creator=40388756"
-	},
-	"item": [
-		{
-			"name": "{{baseUrl}}/auth/oauth2/token",
-			"request": {
-				"auth": {
-					"type": "noauth"
-				},
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n  \"token\": \"y0__xCnyOSxBhilnzgg-7rVuhNAYKLwDD4hsCHFqNBZ-PS3R-4smQ\",\r\n  \"provider\": \"yandex\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "{{baseUrl}}/auth/oauth2/token",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"auth",
-						"oauth2",
-						"token"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "{{baseUrl}}/auth/login",
-			"request": {
-				"auth": {
-					"type": "noauth"
-				},
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n  \"name\": \"Admin\",\r\n  \"email\": \"mistergeki06@gmail.com\",\r\n  \"password\": \"mistergeki123\"\r\n}\r\n",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "{{baseUrl}}/auth/login",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"auth",
-						"login"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "{{baseUrl}}/token/refresh",
-			"request": {
-				"auth": {
-					"type": "noauth"
-				},
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"refreshToken\" : \"eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJtaXN0ZXJnZWtpMDdAZ21haWwuY29tIiwiaWF0IjoxNzQ5NzExNjQ1LCJleHAiOjE3NTIzMDM2NDV9.UCs_tNMiGz1lqDJZ46VzLOanXdG2ZwA4KNc2rszN5xPvTDbAqHGa0bdiux_4Rs4Lz0be-1EbAMw5RdPP7tgxVA\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "{{baseUrl}}/token/refresh",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"token",
-						"refresh"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "logout",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"url": {
-					"raw": "{{baseUrl}}/auth/logout",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"auth",
-						"logout"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
