+++ src/main/java/com/example/marketplace_backend/Model/Brand.java
@@ -22,7 +22,7 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-@Where(clause = "deleted_at IS NULL")
+//@Where(clause = "deleted_at IS NULL")
 public class Brand {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
